KOKU-1296:Tapaamisen peruuttaminen -toiminto kunnan työntekijän sekä kuntalaisen puolelle käyttäjäystävällisemmäksi
Added button for employee view also.
Fixed pop up message position and background color related to if action was succesful.
Green for success and red for failure.
Disabled cancel button while response message is displayed to the user.
Removed cancel button when viewing details of appointment that is already canceled.