KOKULT-4 Portal: disable "cancel" button for non-creator
Also fixed KOKULT-144 - AV event cancellation - comments from internal demo for 1.2.0.
Changed the prompt message to be more user friendly.
Added if statement to check if current (logged in) user is the same as the sender of the appointment in which
case cancel button is displayed. If statement returns false button is hidden from view.