KOKU-1183. KOKULT-152. Added natural sorting for dates.
Added natural sorting to dates from Strings.java.
It's better but not perfect, it fails to sort right in the borders of months but
I'm unable to fix this as long as the data that the webservice gives can include
dates in hour, day or week precision without any type information about which
one it is.