KOKULT-259: Fixed encoding problem with Kunpo and Loora navigation sidebar
There is no Ixonos Jira issue for this fix, because it's a general bug that was
accidentally introduced while working on other features.

Someone's encoding in eclipse was set to UTF-8 in these two files which screwed
up the bytes for some of the ö's and ä's. Would gladly have them as UTF-8 but
seemingly all the production servers are configured to Latin-1 so that's the
encoding that has to be used in files which are displayed directly as web pages
which these two files are. (Source files are obviously unaffected and (at least
for the most part) are UTF-8)