KOKU-1355: Renamed some of the fields in infoRequest detail
KOKULT-388
In information request detail page renamed following things:
	replyDescription: 'Vastauksen lisätiedot' -> 'Vastauksen aihe'
	additionalReplyInfo: 'Vastauksen lisätietoja' -> 'Vastauksen lisätiedot'
	informationDetails: 'Lisätietoja2' -> 'Vastauksen perustelut'

Changed the ordering of the displayed fields in the detail page.

Hid displaying the localizedAccessType in case it is empty. Reasoning being it
is not defined anywhere in the intalio forms nor is it generated in services. So
every entry will have it display empty unless an implementation is made. If such
implementation is made the field will display automatically.